Fixed no reading unused data on ReadT5ileHeader
When reading through a file, ReadT5ileHeader skipped the keys that
where not of interest.
When skipping a key, the whole key data should be read in order to
move the file pointer to the next key and keep synchronization.
Added a missing "read skip_data" if key is not of interest.